Trait pallet_collator_allowlist::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type ValidatorId: Member + Parameter + MaybeSerializeDeserialize + MaxEncodedLen;
    type ValidatorRegistration: ValidatorRegistration<Self::ValidatorId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type WeightInfo: WeightInfo

Type representing the weight of this pallet

type ValidatorId: Member + Parameter + MaybeSerializeDeserialize + MaxEncodedLen

The Validator Id type

type ValidatorRegistration: ValidatorRegistration<Self::ValidatorId>

Type representing the underlying validator registration center. It offers us the API we need to check whether a collator is ready for its duties in the upcoming session.